crate: Extract Schema into its own mod
