+++ src/lib.rs
@@ -1,6 +1,5 @@
 use fnv::FnvHashMap;
 
-use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
 
 pub mod error;
@@ -79,25 +78,8 @@ pub trait GeographicalHeader {
 
 pub mod census2010;
 
-#[derive(Serialize, Deserialize, Copy, Clone, Debug, PartialEq, Eq, Hash)]
-#[non_exhaustive]
-pub enum Schema {
-	Census2010Pl94_171(Option<census2010::pl94_171::Table>),
-}
-
-impl<S: AsRef<str>> core::convert::From<S> for Schema {
-	fn from(s: S) -> Self {
-		let s: &str = s.as_ref();
-		match s {
-			"p1" => Schema::Census2010Pl94_171(Some(census2010::pl94_171::P1)),
-			"p2" => Schema::Census2010Pl94_171(Some(census2010::pl94_171::P2)),
-			"p3" => Schema::Census2010Pl94_171(Some(census2010::pl94_171::P3)),
-			"p4" => Schema::Census2010Pl94_171(Some(census2010::pl94_171::P4)),
-			"h1" => Schema::Census2010Pl94_171(Some(census2010::pl94_171::H1)),
-			_ => unimplemented!(),
-		}
-	}
-}
+mod schema;
+pub use schema::*;
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[non_exhaustive]
@@ -127,27 +109,6 @@ impl FileType {
 		}
 	}
 }
-
-#[cfg(test)]
-mod tests {
-	use crate::census2010::pl94_171::Table;
-	use crate::Schema;
-
-	#[test]
-	fn schema_with_table_de() {
-		let data = r"Census2010Pl94_171: P1";
-		let schema: Schema = serde_yaml::from_str(data).unwrap();
-		assert_eq!(schema, Schema::Census2010Pl94_171(Some(Table::P1)))
-	}
-
-	#[test]
-	fn bare_schema_de() {
-		let data = r"Census2010Pl94_171:";
-		let schema: Schema = serde_yaml::from_str(data).unwrap();
-		assert_eq!(schema, Schema::Census2010Pl94_171(None))
-	}
-}
-
 pub(crate) type GeographicalHeaderIndex = BTreeMap<GeoId, (LogicalRecordNumber, u64)>;
 pub(crate) type LogicalRecordIndex = FnvHashMap<FileType, LogicalRecordPositionIndex>;
